add hashService
